add transaction/stats
delete getBalance
add Balance to current user
