+++ api-doc.json
@@ -16,7 +16,12 @@
     "application/json"
   ],
   "servers": [
-    {}
+    {
+      "url": "https://wallet-proj.herokuapp.com/"
+    },
+    {
+      "url": "http://localhost:4000/"
+    }
   ],
   "tags": [
     {
@@ -212,48 +217,6 @@
         }
       }
     },
-    "/api/users/balance": {
-      "get": {
-        "tags": [
-          "Users"
-        ],
-        "summary": "get user balance (requires authentication token)",
-        "parameters": [],
-        "security": [
-          {
-            "Bearer": []
-          }
-        ],
-        "responses": {
-          "200": {
-            "description": "Returns user information",
-            "content": {
-              "application/json": {
-                "schema": {
-                  "$ref": "#/components/schemas/User-balance-response"
-                }
-              }
-            }
-          },
-          "401": {
-            "description": "User unauthorized",
-            "content": {
-              "application/json": {
-                "schema": {
-                  "type": "object",
-                  "properties": {
-                    "message": {
-                      "type": "string",
-                      "example": "Not authorized"
-                    }
-                  }
-                }
-              }
-            }
-          }
-        }
-      }
-    },
     "/api/categories": {
       "get": {
         "tags": [
@@ -345,6 +308,46 @@
               "application/json": {
                 "schema": {
                   "$ref": "#/components/schemas/Transactions-response"
+                }
+              }
+            }
+          },
+          "401": {
+            "description": "Unauthorized"
+          }
+        }
+      }
+    },
+    "/api/transactions/stats": {
+      "post": {
+        "tags": [
+          "Transactions"
+        ],
+        "summary": "User's transactions statistic",
+        "parameters": [],
+        "security": [
+          {
+            "Bearer": []
+          }
+        ],
+        "requestBody": {
+          "description": "Body of request for user statistic",
+          "required": true,
+          "content": {
+            "application/json": {
+              "schema": {
+                "$ref": "#/components/schemas/Transactions-stats-request"
+              }
+            }
+          }
+        },
+        "responses": {
+          "200": {
+            "description": "User's Transactions statistics",
+            "content": {
+              "application/json": {
+                "schema": {
+                  "$ref": "#/components/schemas/Transactions-stats-response"
                 }
               }
             }
@@ -467,23 +470,10 @@
           "name": {
             "type": "string",
             "example": "Mango"
-          }
-        }
-      },
-      "User-balance-response": {
-        "type": "object",
-        "properties": {
-          "email": {
-            "type": "string",
-            "example": "mango@gmail.com"
-          },
-          "name": {
-            "type": "string",
-            "example": "Mango"
           },
           "balance": {
-            "type":"number",
-          "example": 0.00
+            "type": "number",
+            "example": 0.00
           }
         }
       },
@@ -602,6 +592,57 @@
             },
             "createdAt": "2022-04-04T07:48:24.034Z",
             "updatedAt": "2022-04-04T07:48:24.034Z"
+          }
+        ]
+      },
+      "Transactions-stats-request": {
+        "type": "object",
+        "required": [
+          "begin",
+          "end"
+        ],
+        "properties": {
+          "begin": {
+            "type": "string",
+            "description": "Date of period begin",
+            "example": "01/01/2022"
+          },
+          "end": {
+            "type": "string",
+            "description": "Date of period end",
+            "example": "12/01/2022"
+          }
+        }
+      },
+      "Transactions-stats-response": {
+        "type": "array",
+        "items": {
+          "type": "object",
+          "properties": {
+            "category": {
+              "type": "string",
+              "description": "Category",
+              "example": "add"
+            },
+            "totalSum": {
+              "type": "number",
+              "description": "Total sum of the current category",
+              "example": 0
+            }
+          }
+        },
+        "example": [
+          {
+            "category": "add2",
+            "totalSum": 3600
+          },
+          {
+            "category": "add8",
+            "totalSum": 2000
+          },
+          {
+            "category": "add15",
+            "totalSum": 1600
           }
         ]
       }
