+++ controllers/transactions/index.js
@@ -1,7 +1,9 @@
 const listTransactions = require('./listTransaction');
+const statsTransactions = require('./statsTransaction');
 const addTransaction = require('./addTransaction');
 
 module.exports = {
   listTransactions,
+  statsTransactions,
   addTransaction,
 };
