+++ controllers/users/index.js
@@ -3,7 +3,7 @@ const login = require('./login');
 // const verify = require('./verify');
 // const repeatVerify = require('./repeatVerify');
 const getCurrent = require('./getCurrent');
-const getBalance = require('./getBalance');
+
 const logout = require('./logout');
 
 module.exports = {
@@ -12,6 +12,6 @@ module.exports = {
   // verify,
   // repeatVerify,
   getCurrent,
-  getBalance,
+
   logout,
 };
