+++ routes/api/transactions.js
@@ -10,6 +10,8 @@ router.get('/', auth, ctrlWrapper(ctrl.listTransactions));
 
 router.post('/', [auth, validation(schemas.add)], ctrlWrapper(ctrl.addTransaction));
 
+router.post('/stats', auth, ctrlWrapper(ctrl.statsTransactions));
+
 // router.get('/:contactId', auth, ctrlWrapper(ctrl.getContactById));
 
 // router.delete('/:transactionId', auth, ctrlWrapper(ctrl.removeTransaction));
