+++ routes/api/users.js
@@ -6,6 +6,5 @@ const { users: ctrl } = require('../../controllers');
 const router = express.Router();
 
 router.get('/current', auth, ctrlWrapper(ctrl.getCurrent));
-router.get('/balance', auth, ctrlWrapper(ctrl.getBalance));
 
 module.exports = router;
